refactor: :recycle: update drinks dto
