+++ src/modules/drinks/dtos/ingredients.ts
@@ -1,0 +1,11 @@
+interface IIngredient {
+	id?: string;
+	name: string;
+	unity: string;
+	category: string;
+	isAlcoholic: boolean;
+	colorTheme: string;
+	created_at?: Date;
+}
+
+export { IIngredient };
