+++ src/modules/drinks/infra/prisma/DrinksRepository.ts
@@ -1,7 +1,7 @@
 import { PrismaClient, Prisma } from '@prisma/client';
 import { getPrismaClient } from '@shared/container/providers/prisma';
 import { IDrinksRepository } from '@modules/drinks/repositories/IDrinksRepository';
-import { IDrinkResponse, IDrink } from '@modules/drinks/dtos/DrinksDTO';
+import { IDrinkResponse, IDrink } from '@modules/drinks/dtos/Drinks';
 
 interface IDrinkAggregation {
 	_id: {
